Prefer system Abseil if available
Newer versions of Protobuf (22+) pull in Abseil as a dependency. We
want to avoid using our bundled copy for these cases, as this will
likely conflict with the version of Abseil that Protobuf uses.

The simplest way to do this seems to be to just prefer a system
installation of Abseil if it's available.

Signed-off-by: Carlo Cabrera <30379873+carlocab@users.noreply.github.com>